+++ storybook/lib/stories/segmented_control.dart
@@ -6,18 +6,50 @@ import 'package:storybook_flutter/storybook_flutter.dart';
 
 final Story segmentedControlStory = Story(
   name: 'Segmented control',
-  // builder: (_, k) => _SegmentedControlExample(
-  //   label: k.text(label: 'Label', initial: ''),
-  //   error: k.text(label: 'Error', initial: ''),
-  //   isEnabled: k.boolean(label: 'Enabled', initial: true),
-  //   isRequired: k.boolean(label: 'Required', initial: false),
-  // ),
-  builder: (_, k) => _SegmentedControlExample(
-    label: 'Label',
-    error:'Error',
-    isEnabled: true,
-    isRequired: true,
-  ),
+  builder: (_, k) {
+    final label = k.text(label: 'Label', initial: '');
+    final error = k.text(label: 'Error', initial: '');
+    final isEnabled = k.boolean(label: 'Enabled', initial: true);
+    final isRequired = k.boolean(label: 'Required', initial: false);
+
+    return SingleChildScrollView(
+      child: Column(
+        mainAxisSize: MainAxisSize.min,
+        children: [
+          Padding(
+            padding: const EdgeInsets.symmetric(vertical: spacing100),
+            child: _SegmentedControlExample(
+              label: label,
+              error: error,
+              isEnabled: isEnabled,
+              isRequired: isRequired,
+              options: _options2,
+            ),
+          ),
+          Padding(
+            padding: const EdgeInsets.symmetric(vertical: spacing100),
+            child: _SegmentedControlExample(
+              label: label,
+              error: error,
+              isEnabled: isEnabled,
+              isRequired: isRequired,
+              options: _options3,
+            ),
+          ),
+          Padding(
+            padding: const EdgeInsets.symmetric(vertical: spacing100),
+            child: _SegmentedControlExample(
+              label: label,
+              error: error,
+              isEnabled: isEnabled,
+              isRequired: isRequired,
+              options: _options4,
+            ),
+          ),
+        ],
+      ),
+    );
+  },
 );
 
 class _SegmentedControlExample extends StatefulWidget {
@@ -27,12 +59,14 @@ class _SegmentedControlExample extends StatefulWidget {
     required this.error,
     required this.isEnabled,
     required this.isRequired,
+    required this.options,
   }) : super(key: key);
 
   final String label;
   final String error;
   final bool isEnabled;
   final bool isRequired;
+  final List<String> options;
 
   @override
   _SegmentedControlExampleState createState() =>
@@ -50,15 +84,26 @@ class _SegmentedControlExampleState extends State<_SegmentedControlExample> {
         error: widget.error,
         isEnabled: widget.isEnabled,
         onChanged: (value) => setState(() => _groupValue = value),
-        items: _options
+        items: widget.options
             .map((i) => OptimusGroupItem<String>(label: Text(i), value: i))
             .toList(),
       );
 }
 
-const _options = [
+const _options2 = [
   'Another long option',
-  'Bxcxc',
-  'Cxcxcxc',
-  'Dxcxcxc',
+  'B',
+];
+
+const _options3 = [
+  'Another long option',
+  'B',
+  'C',
+];
+
+const _options4 = [
+  'Another long option',
+  'B',
+  'C',
+  'D',
 ];
